move: 패키지 정리 및 import 제거
ExpressiongValidator클래스 domain 패키지로 이동, test 패키지와 main 패키지 구조를 동일하게 변경
