+++ src/test/java/com/programmers/domain/ExpressionValidatorTest.java
@@ -1,6 +1,5 @@
-package com.programmers.validator;
+package com.programmers.domain;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
