+++ src/test/java/com/programmers/repository/CalculationMemoryRepositoryTest.java
@@ -1,12 +1,10 @@
 package com.programmers.repository;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class CalculationMemoryRepositoryTest {
     CalculationRepository calculationRepository = new CalculationMemoryRepository();
